Updated Project

+++ src/main/java/com/facens/pooii/lab/ac1/ac1/services/EventService.java
@@ -1,5 +1,7 @@
 package com.facens.pooii.lab.ac1.ac1.services;
 
+import java.io.Console;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -8,13 +10,24 @@ import javax.persistence.EntityNotFoundException;
 
 import com.facens.pooii.lab.ac1.ac1.dtos.EventDTO;
 import com.facens.pooii.lab.ac1.ac1.dtos.EventInsertDTO;
+import com.facens.pooii.lab.ac1.ac1.dtos.EventTicketSellDTO;
 import com.facens.pooii.lab.ac1.ac1.dtos.EventUpdateDTO;
+import com.facens.pooii.lab.ac1.ac1.dtos.TicketDTO;
 import com.facens.pooii.lab.ac1.ac1.entities.Admin;
+import com.facens.pooii.lab.ac1.ac1.entities.Attend;
 import com.facens.pooii.lab.ac1.ac1.entities.Event;
+import com.facens.pooii.lab.ac1.ac1.entities.Place;
+import com.facens.pooii.lab.ac1.ac1.entities.Ticket;
 import com.facens.pooii.lab.ac1.ac1.repositories.AdminRepository;
+import com.facens.pooii.lab.ac1.ac1.repositories.AttendRepository;
 import com.facens.pooii.lab.ac1.ac1.repositories.EventRepository;
+import com.facens.pooii.lab.ac1.ac1.repositories.PlaceRepository;
+import com.facens.pooii.lab.ac1.ac1.repositories.TicketRepository;
 import com.facens.pooii.lab.ac1.ac1.utils.DateTimeValidation;
 import com.facens.pooii.lab.ac1.ac1.utils.FilterEventRequest;
+import com.facens.pooii.lab.ac1.ac1.utils.PlaceEventValidation;
+import com.facens.pooii.lab.ac1.ac1.utils.SellTicketValidation;
+import com.facens.pooii.lab.ac1.ac1.utils.TicketType.ticketType;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -31,6 +44,12 @@ public class EventService {
     private EventRepository repo;
     @Autowired
     private AdminRepository repoAdmin;
+    @Autowired
+    private PlaceRepository repoPlace;
+    @Autowired
+    private AttendRepository repoAttend;
+    @Autowired
+    private TicketRepository repoTicket;
 
     public Page<EventDTO> getEvents(PageRequest pageRequest, FilterEventRequest filterRequest){
         Page<Event> list = repo.find(pageRequest, filterRequest.getName(), filterRequest.getStartDate(), filterRequest.getDescription(), filterRequest.getPriceTicket());
@@ -58,6 +77,15 @@ public class EventService {
     }
 
     public void delete(Long id){
+        Optional<Event> opEvent = repo.findById(id);
+        Event event = opEvent.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+
+        if(event.getPayedTicketsSelled() > 0 || event.getFreeTicketsSelled() > 0){
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED, "This event have tickets with Attendees!!!!"
+            );
+        }
+
         try{
             repo.deleteById(id);
         }catch(EmptyResultDataAccessException e){
@@ -99,11 +127,168 @@ public class EventService {
                 e.getEmailContact(),
                 e.getAmountFreeTickets(),
                 e.getAmountPayedTickets(),
-                e.getPriceTicket()
+                e.getPriceTicket(),
+                e.getFreeTicketsSelled(),
+                e.getPayedTicketsSelled(),
+                e.getPlaces(), 
+                e.getTickets()
             );
             listDTO.add(dto);
         }
 
         return listDTO;
+    }
+
+    public EventDTO addLocalEvent(Long idEvent, Long idPlace) {
+        Optional<Event> opEvent = repo.findById(idEvent);
+        Event event = opEvent.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+
+        Optional<Place> opPlace = repoPlace.findById(idPlace);
+        Place place = opPlace.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Place not found"));
+
+        PlaceEventValidation.valide(event, place);
+
+        event.addPlace(place);
+        repo.save(event);
+
+        place.addEvent(event);
+        repoPlace.save(place);       
+
+        return new EventDTO(event);
+    }
+
+    public EventDTO removeLocalEvent(Long idEvent, Long idPlace) {
+        Optional<Event> opEvent = repo.findById(idEvent);
+        Event event = opEvent.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+
+        Optional<Place> opPlace = repoPlace.findById(idPlace);
+        Place place = opPlace.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Place not found"));
+
+        if(event.getPlaces().contains(place)){
+            event.getPlaces().remove(place);
+            place.getEvents().remove(event);
+        } else {
+            throw new ResponseStatusException(
+                    HttpStatus.UNAUTHORIZED, "The place isn't registered in the event!"
+            );
+        }
+
+        repo.save(event);
+        repoPlace.save(place);       
+
+        return new EventDTO(event);
+    }
+
+    public EventDTO sellTicket(EventTicketSellDTO dto, Long idEvent){
+        Optional<Attend> opAttend = repoAttend.findById(dto.getIdAttend());
+        Attend attend = opAttend.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Attend not found"));
+
+        Optional<Event> opEvent = repo.findById(idEvent);
+        Event event = opEvent.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+
+        ticketType ticketT = SellTicketValidation.valide(dto.getTypeTicket());
+
+        switch(ticketT){
+            case FREE:
+                if(event.getAmountFreeTickets() > 0){
+                    Ticket ticket = new Ticket(event, attend, ticketT, 0d, Instant.now());
+                    for(Ticket atTicket : attend.getTickets())
+                        if(atTicket.getEvent().getId() == ticket.getEvent().getId())
+                            throw new ResponseStatusException(
+                                HttpStatus.UNAUTHORIZED, "This attend already have this ticket!"
+                            );
+                    repoTicket.save(ticket);
+
+                    event.setAmountFreeTickets(event.getAmountFreeTickets()-1);
+                    event.setFreeTicketsSelled(event.getFreeTicketsSelled()+1);
+                    event.addTicket(ticket);
+                    repo.save(event);
+
+                    attend.addTicket(ticket);
+                    repoAttend.save(attend);                    
+                } else
+                    throw new ResponseStatusException(
+                        HttpStatus.UNAUTHORIZED, "Doest have available FREE tickets!"
+                    );
+                break;
+            case PAYED:
+                if(event.getAmountPayedTickets() > 0){
+                    Ticket ticket = new Ticket(event, attend, ticketT, event.getPriceTicket(), Instant.now());
+                    for(Ticket atTicket : attend.getTickets())
+                        if(atTicket.getEvent().getId() == ticket.getEvent().getId())
+                            throw new ResponseStatusException(
+                                HttpStatus.UNAUTHORIZED, "This attend already have this ticket!"
+                            );
+
+                    repoTicket.save(ticket);
+
+                    event.setAmountPayedTickets(event.getAmountPayedTickets()-1);
+                    event.setPayedTicketsSelled(event.getPayedTicketsSelled()+1);
+                    event.addTicket(ticket);
+                    repo.save(event);
+
+                    attend.addTicket(ticket);
+                    repoAttend.save(attend);  
+
+                    event.setAmountPayedTickets(event.getAmountPayedTickets()-1);
+                } else
+                    throw new ResponseStatusException(
+                        HttpStatus.UNAUTHORIZED, "Doest have available PAYED tickets!"
+                    );
+                break;
+        }
+        
+        return new EventDTO(event);
+    }
+
+    public EventDTO devolutionTicket(EventTicketSellDTO dto, Long idEvent) {
+        Optional<Attend> opAttend = repoAttend.findById(dto.getIdAttend());
+        Attend attend = opAttend.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Attend not found"));
+
+        Optional<Event> opEvent = repo.findById(idEvent);
+        Event event = opEvent.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+
+        ticketType ticketT = SellTicketValidation.valide(dto.getTypeTicket());
+
+        for(Ticket atTicket : attend.getTickets())
+            if(atTicket.getEvent().getId() == idEvent && atTicket.getType() == ticketT){
+                switch(ticketT){
+                    case FREE:
+                        event.setAmountFreeTickets(event.getAmountFreeTickets()+1);
+                        event.setFreeTicketsSelled(event.getFreeTicketsSelled()-1);
+                        event.removeTicket(atTicket);
+                        repo.save(event);
+
+                        attend.removeTicket(atTicket);
+                        repoAttend.save(attend);
+
+                        repoTicket.delete(atTicket);
+                        return new EventDTO(event);
+
+                    case PAYED:
+                        event.setAmountPayedTickets(event.getAmountPayedTickets()+1);
+                        event.setPayedTicketsSelled(event.getPayedTicketsSelled()-1);
+                        event.removeTicket(atTicket);
+                        repo.save(event);
+
+                        attend.setBalance(attend.getBalance()+event.getPriceTicket());
+                        attend.removeTicket(atTicket);
+                        repoAttend.save(attend);
+
+                        repoTicket.delete(atTicket);
+                        return new EventDTO(event);
+                }
+            }
+
+        throw new ResponseStatusException(
+            HttpStatus.UNAUTHORIZED, "The attend doesn't have ticket for this event!"
+        );
+    }
+
+    public TicketDTO getTickets(Long idEvent) {
+        Optional<Event> opEvent = repo.findById(idEvent);
+        Event event = opEvent.orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+
+        return new TicketDTO(event);
     }
 }
